[tls] Allow clients to authenticate with servers using a certificate
Servers may send a `CertificateRequest` to clients to indicate they wish
the client to authenticate using a certificate. We add support for this
on the client-side, allowing clients to perform the handshake,
regardless of whether they do or don't have a certificate.

The server-side code is mostly there for testing purposes as there is
currently no mechanism to allow server API consumers to request a client
certificate or validate it.

Fixes: #377