[h3] Adapt datagram API for RFC 9297 (fixes: #420)
Our HTTP/3 code dealing with datagrams dates back to before RFC 9297 was
published. This is reflected in the the use of the `flow_id` terminology
which no longer exists. Update the following interfaces:

- `DatagramReceived` now has a `stream_id` attribute
  instead of `flow_id`.
- `H3Connection.send_datagram` now has a `stream_id` argument
  instead of `flow_id`

Encoding and decoding to / from quarter stream ID is handled by the
connection, so the user does not need to multiply or divide by four.