Explicitly track whether FIN has been acknowledged (fixes: #438)
A stream's sender part is only finished if all the data and the FIN have
been acknowledged. Tracking the acknowledged data ranges is not
sufficient to know whether the FIN has been acknowledged too, as the FIN
may have been sent on its own. To fix this we need to explicitly keep
track of whether the frame containing the FIN was acknowledged.

We also:

- Remove a bogus assignment of `QuicStreamSender.send_buffer_empty`,
  this should have been `QuicStreamSender.buffer_is_empty`.
- Only set `buffer_is_empty` to `False` when data or a FIN have been
  lost. This saves useless calls to `QuicStreamSender.get_frame()`.

Co-authored-by: Maximilian Hils <git@maximilianhils.com>