Do not fully clear ACK ranges when ACK is acknowledged
When a packet containing an ACK is acknowledged, we are allowed to
discard ACK ranges up to and including the highest acknowledged packet
in the sent ACK:

https://datatracker.ietf.org/doc/html/rfc9000#ack-tracking

However, doing so prevents us from reporting a gap if the immediately
following packet(s) are lost. We therefore retain at least the highest
acknowledged packet in our ACK ranges.

Fixes: #495.