During address validation, count the entire received datagram
For anti-amplification purposes, servers need to keep track of the
amount of data received on unvalidated network paths. We must count the
entire datagram size regardless of whether packets are processed or
dropped.

This is particularly important when talking to clients who pad
datagrams containing INITIAL packets by appending bytes after the
long-header packets, which is legitimate behaviour.