h3: Remove stream objects when stream is ended
The H3Connection._stream dict is never cleared of created streams,
leading to a memory leak.

This commit addresses the issue. First, it introduces tracking of
whether HTTP/3 streams have ended on both sides. Then, it refactors
H3Connection._get_or_create_stream to be a context manager which
automatically removes streams if they are ended on both sides and
are not blocked. Finally, it adds a test for this case, though it's
a bit dirty because it accesses internal H3Connection field, but it
works well overall.